Externalize GET schemas
